fix: dont call MacCssFix constantly
